backend/pyproject.toml
[project]
name = "atlas-historico-back"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "geoalchemy2>=0.18.1",
    "geopy>=2.4.1",
    "kagglehub>=0.3.13",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "requests>=2.32.5",
    "shapely>=2.1.2",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
]
